+++ src/components/Button/Button.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import { Button, ButtonProps } from "./bugButton";
+import { Button, ButtonProps } from "./Button";
 import { ButtonVariant } from "../../foundation/Button";
 
 const meta: Meta<ButtonProps> = {
