+++ src/components/Button/index.ts
@@ -1,1 +1,1 @@
-export * from "./bugButton";
+export * from "./Button";
